Add error boundary test
